+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.2.1-dev
 - Added a new `HTTP_CACHE_TIME` setting to control caching of static web files. Defaults to 0 for easier development, which means that all static files (HTML, JS, CSS...) are not cached by web browsers.
+- The list of all loaded endpoints is now displayed when using `silence new`. Added a new `SHOW_ENDPOINT_LIST` setting to disable this behavior.
 
 # 1.2.1
 - Added `cryptography` to dependencies for compatibility with MySQL.